+++ CHANGELOG.md
@@ -138,6 +138,7 @@
   `DQLQueryBuilderInterface::getExpressionBuilder()` methods to `ExpressionBuilderInterface` (@vjik)
 - New #1029: Add functions as expressions (@Tigrov)
 - Enh #1042: Refactor `AbstractDMLQueryBuilder` class to `upsert()` method (@Tigrov)
+- New #1040: Add `DateTimeValue` class (@vjik)
 
 ## 1.3.0 March 21, 2024